@@ -5,9 +5,14 @@ import 'package:flutter_reactive_ble_example/src/ble/reactive_state.dart';
 import 'package:meta/meta.dart';
 
 class BleScanner implements ReactiveState<BleScannerState> {
-  BleScanner(this._ble);
+  BleScanner({
+    required FlutterReactiveBle ble,
+    required Function(String message) logMessage,
+  })   : _ble = ble,
+        _logMessage = logMessage;
 
   final FlutterReactiveBle _ble;
+  final void Function(String message) _logMessage;
   final StreamController<BleScannerState> _stateStreamController =
       StreamController();
 
@@ -17,6 +22,7 @@ class BleScanner implements ReactiveState<BleScannerState> {
   Stream<BleScannerState> get state => _stateStreamController.stream;
 
   void startScan(List<Uuid> serviceIds) {
+    _logMessage('Start ble discovery');
     _devices.clear();
     _subscription?.cancel();
     _subscription =
@@ -28,7 +34,7 @@ class BleScanner implements ReactiveState<BleScannerState> {
         _devices.add(device);
       }
       _pushState();
-    });
+    }, onError: (e) => _logMessage('Device scan fails with error: $e'));
     _pushState();
   }
 
@@ -42,6 +48,8 @@ class BleScanner implements ReactiveState<BleScannerState> {
   }
 
   Future<void> stopScan() async {
+    _logMessage('Stop ble discovery');
+
     await _subscription?.cancel();
     _subscription = null;
     _pushState();
@@ -51,14 +59,14 @@ class BleScanner implements ReactiveState<BleScannerState> {
     await _stateStreamController.close();
   }
 
-  StreamSubscription _subscription;
+  StreamSubscription? _subscription;
 }
 
 @immutable
 class BleScannerState {
   const BleScannerState({
-    @required this.discoveredDevices,
-    @required this.scanIsInProgress,
+    required this.discoveredDevices,
+    required this.scanIsInProgress,
   });
 
   final List<DiscoveredDevice> discoveredDevices;
