@@ -9,12 +9,12 @@ abstract class ConnectedDeviceOperation {
 
   Future<void> writeCharacteristicWithResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   });
 
   Future<void> writeCharacteristicWithoutResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   });
 
   Stream<List<int>> subscribeToCharacteristic(
@@ -32,9 +32,8 @@ abstract class ConnectedDeviceOperation {
 
 class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   ConnectedDeviceOperationImpl({
-    @required DeviceOperationController controller,
-  })  : assert(controller != null),
-        _controller = controller;
+    required DeviceOperationController controller,
+  }) : _controller = controller;
 
   final DeviceOperationController _controller;
 
@@ -50,7 +49,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
 
     return _controller
         .readCharacteristic(characteristic)
-        .asyncExpand((Object _) => specificCharacteristicValueStream)
+        .asyncExpand((_) => specificCharacteristicValueStream)
         .firstWhere((_) => true,
             orElse: () => throw NoBleCharacteristicDataReceived());
   }
@@ -58,7 +57,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   @override
   Future<void> writeCharacteristicWithResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   }) async =>
       _controller
           .writeCharacteristicWithResponse(characteristic, value)
@@ -67,7 +66,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   @override
   Future<void> writeCharacteristicWithoutResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   }) async =>
       _controller
           .writeCharacteristicWithoutResponse(characteristic, value)
@@ -85,7 +84,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
     final autosubscribingRepeater = Repeater<List<int>>.broadcast(
       onListenEmitFrom: () => _controller
           .subscribeToNotifications(characteristic)
-          .asyncExpand((Object _) => specificCharacteristicValueStream),
+          .asyncExpand((_) => specificCharacteristicValueStream),
       onCancel: () => _controller
           .stopSubscribingToNotifications(characteristic)
           .catchError((Object e) =>
