@@ -4,7 +4,6 @@ import 'package:flutter_reactive_ble/src/converter/args_to_protubuf_converter.da
 import 'package:flutter_reactive_ble/src/debug_logger.dart';
 import 'package:flutter_reactive_ble/src/model/uuid.dart';
 import 'package:flutter_reactive_ble/src/model/write_characteristic_info.dart';
-import 'package:meta/meta.dart';
 
 import 'converter/protobuf_converter.dart';
 import 'model/characteristic_value.dart';
@@ -39,7 +38,7 @@ abstract class DeviceOperationController {
     QualifiedCharacteristic characteristic,
   );
 
-  Future<int> requestMtuSize(String deviceId, int mtu);
+  Future<int> requestMtuSize(String deviceId, int? mtu);
 
   Future<ConnectionPriorityInfo> requestConnectionPriority(
       String deviceId, ConnectionPriority priority);
@@ -49,9 +48,9 @@ abstract class ScanOperationController {
   Stream<ScanResult> get scanStream;
 
   Stream<void> scanForDevices({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid> withServices,
+    required ScanMode scanMode,
+    required bool requireLocationServicesEnabled,
   });
 }
 
@@ -60,8 +59,8 @@ abstract class DeviceConnectionController {
 
   Stream<void> connectToDevice(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   );
 
   Future<void> disconnectDevice(String deviceId);
@@ -74,7 +73,7 @@ abstract class BleOperationController {
 
   Future<void> deinitialize();
 
-  Future<Result<Unit, GenericFailure<ClearGattCacheError>>> clearGattCache(
+  Future<Result<Unit, GenericFailure<ClearGattCacheError>?>> clearGattCache(
       String deviceId);
 }
 
@@ -85,23 +84,15 @@ class PluginController
         DeviceConnectionController,
         BleOperationController {
   PluginController({
-    @required ArgsToProtobufConverter argsToProtobufConverter,
-    @required ProtobufConverter protobufConverter,
-    @required MethodChannel bleMethodChannel,
-    @required Stream<List<int>> connectedDeviceChannel,
-    @required Stream<List<int>> charUpdateChannel,
-    @required Stream<List<int>> bleDeviceScanChannel,
-    @required Stream<List<int>> bleStatusChannel,
-    @required Logger debugLogger,
-  })  : assert(argsToProtobufConverter != null),
-        assert(protobufConverter != null),
-        assert(bleMethodChannel != null),
-        assert(connectedDeviceChannel != null),
-        assert(bleDeviceScanChannel != null),
-        assert(bleStatusChannel != null),
-        assert(charUpdateChannel != null),
-        assert(debugLogger != null),
-        _argsToProtobufConverter = argsToProtobufConverter,
+    required ArgsToProtobufConverter argsToProtobufConverter,
+    required ProtobufConverter protobufConverter,
+    required MethodChannel bleMethodChannel,
+    required Stream<List<int>> connectedDeviceChannel,
+    required Stream<List<int>> charUpdateChannel,
+    required Stream<List<int>> bleDeviceScanChannel,
+    required Stream<List<int>> bleStatusChannel,
+    required Logger debugLogger,
+  })   : _argsToProtobufConverter = argsToProtobufConverter,
         _protobufConverter = protobufConverter,
         _bleMethodChannel = bleMethodChannel,
         _connectedDeviceRawStream = connectedDeviceChannel,
@@ -119,10 +110,10 @@ class PluginController
   final Stream<List<int>> _bleStatusRawChannel;
   final Logger _debugLogger;
 
-  Stream<ConnectionStateUpdate> _connectionUpdateStream;
-  Stream<CharacteristicValue> _charValueStream;
-  Stream<ScanResult> _scanResultStream;
-  Stream<BleStatus> _bleStatusStream;
+  Stream<ConnectionStateUpdate>? _connectionUpdateStream;
+  Stream<CharacteristicValue>? _charValueStream;
+  Stream<ScanResult>? _scanResultStream;
+  Stream<BleStatus>? _bleStatusStream;
 
   @override
   Stream<ConnectionStateUpdate> get connectionUpdateStream =>
@@ -182,9 +173,9 @@ class PluginController
 
   @override
   Stream<void> scanForDevices({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid> withServices,
+    required ScanMode scanMode,
+    required bool requireLocationServicesEnabled,
   }) {
     _debugLogger.log(
       'Start scanning for devices with arguments (withServices:$withServices, scanMode: $scanMode, locationServiceEnabled: $requireLocationServicesEnabled)',
@@ -206,8 +197,8 @@ class PluginController
   @override
   Stream<void> connectToDevice(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   ) {
     _debugLogger.log(
       'Start connecting to device with arguments (deviceId: $id, servicesWithCharacteristicsToDiscover: $servicesWithCharacteristicsToDiscover, timeout: $connectionTimeout)',
@@ -262,7 +253,7 @@ class PluginController
             _argsToProtobufConverter
                 .createWriteChacracteristicRequest(characteristic, value)
                 .writeToBuffer())
-        .then(_protobufConverter.writeCharacteristicInfoFrom);
+        .then((data) => _protobufConverter.writeCharacteristicInfoFrom(data!));
   }
 
   @override
@@ -279,7 +270,7 @@ class PluginController
               .createWriteChacracteristicRequest(characteristic, value)
               .writeToBuffer(),
         )
-        .then(_protobufConverter.writeCharacteristicInfoFrom);
+        .then((data) => _protobufConverter.writeCharacteristicInfoFrom(data!));
   }
 
   @override
@@ -316,17 +307,17 @@ class PluginController
   }
 
   @override
-  Future<int> requestMtuSize(String deviceId, int mtu) async {
+  Future<int> requestMtuSize(String deviceId, int? mtu) async {
     _debugLogger
         .log('Request mtu size for device: $deviceId with mtuSize: $mtu');
     return _bleMethodChannel
         .invokeMethod<List<int>>(
           "negotiateMtuSize",
           _argsToProtobufConverter
-              .createNegotiateMtuRequest(deviceId, mtu)
+              .createNegotiateMtuRequest(deviceId, mtu!)
               .writeToBuffer(),
         )
-        .then(_protobufConverter.mtuSizeFrom);
+        .then((data) => _protobufConverter.mtuSizeFrom(data!));
   }
 
   @override
@@ -341,11 +332,11 @@ class PluginController
               .createChangeConnectionPrioRequest(deviceId, priority)
               .writeToBuffer(),
         )
-        .then(_protobufConverter.connectionPriorityInfoFrom);
+        .then((data) => _protobufConverter.connectionPriorityInfoFrom(data!));
   }
 
   @override
-  Future<Result<Unit, GenericFailure<ClearGattCacheError>>> clearGattCache(
+  Future<Result<Unit, GenericFailure<ClearGattCacheError>?>> clearGattCache(
       String deviceId) {
     _debugLogger.log('Clear gatt cache for device: $deviceId');
     return _bleMethodChannel
@@ -355,7 +346,7 @@ class PluginController
               .createClearGattCacheRequest(deviceId)
               .writeToBuffer(),
         )
-        .then(_protobufConverter.clearGattCacheResultFrom);
+        .then((data) => _protobufConverter.clearGattCacheResultFrom(data!));
   }
 
   @override
@@ -367,7 +358,7 @@ class PluginController
                 .createDiscoverServicesRequest(deviceId)
                 .writeToBuffer(),
           )
-          .then(_protobufConverter.discoveredServicesFrom);
+          .then((data) => _protobufConverter.discoveredServicesFrom(data!));
 }
 
 class PluginControllerFactory {
