@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:flutter_reactive_ble/src/connected_device_operation.dart';
@@ -25,14 +26,14 @@ import 'reactive_ble_test.mocks.dart';
 )
 void main() {
   group('$FlutterReactiveBle', () {
-    MockBleOperationController _bleOperationController;
-    MockDeviceScanner _deviceScanner;
-    MockDeviceConnector _deviceConnector;
-    MockConnectedDeviceOperation _deviceOperation;
-    StreamController<BleStatus> _bleStatusController;
+    late MockBleOperationController _bleOperationController;
+    late MockDeviceScanner _deviceScanner;
+    late MockDeviceConnector _deviceConnector;
+    late MockConnectedDeviceOperation _deviceOperation;
+    late StreamController<BleStatus> _bleStatusController;
     MockLogger _debugLogger;
 
-    FlutterReactiveBle _sut;
+    late FlutterReactiveBle _sut;
 
     setUp(() {
       _bleOperationController = MockBleOperationController();
@@ -45,6 +46,11 @@ void main() {
       when(_bleOperationController.initialize()).thenAnswer(
         (_) => Future.value(),
       );
+
+      when(_bleOperationController.deinitialize()).thenAnswer(
+        (_) => Future.value(),
+      );
+
       when(_bleOperationController.bleStatusStream).thenAnswer(
           (realInvocation) => _bleStatusController.stream.asBroadcastStream());
 
@@ -54,15 +60,17 @@ void main() {
         deviceConnector: _deviceConnector,
         connectedDeviceOperation: _deviceOperation,
         debugLogger: _debugLogger,
+        initialization: Future.value(),
       );
     });
 
     tearDown(() async {
+      await _sut.deinitialize();
       await _bleStatusController.close();
     });
 
     group('BleStatus stream', () {
-      Stream<BleStatus> bleStatusStream;
+      Stream<BleStatus>? bleStatusStream;
       setUp(() {
         bleStatusStream = _sut.statusStream;
       });
@@ -99,7 +107,7 @@ void main() {
       );
       final charValue = CharacteristicValue(
           characteristic: _createChar(), result: characteristic);
-      Stream<CharacteristicValue> charValueStream;
+      Stream<CharacteristicValue>? charValueStream;
 
       setUp(() {
         when(_deviceOperation.characteristicValueStream)
@@ -130,7 +138,7 @@ void main() {
 
     group('Read characteristic', () {
       QualifiedCharacteristic characteristic;
-      List<int> result;
+      List<int>? result;
 
       setUp(() async {
         characteristic = _createChar();
@@ -192,7 +200,7 @@ void main() {
     group('Request mtu', () {
       const deviceId = '123';
       const mtu = 120;
-      int result;
+      int? result;
 
       setUp(() async {
         when(_deviceOperation.requestMtu(any, any))
@@ -227,14 +235,14 @@ void main() {
       const mode = ScanMode.lowPower;
       const requireLocation = false;
 
-      const device = DiscoveredDevice(
+      final device = DiscoveredDevice(
         id: 'deviceId',
-        manufacturerData: null,
+        manufacturerData: Uint8List.fromList([0]),
         name: 'test',
         rssi: -39,
         serviceData: {},
       );
-      Stream<DiscoveredDevice> deviceStream;
+      Stream<DiscoveredDevice>? deviceStream;
 
       setUp(() {
         when(_deviceScanner.scanForDevices(
@@ -273,7 +281,7 @@ void main() {
       final servicesToDiscover = {
         Uuid.parse('FEFF'): [Uuid.parse('FE1F')]
       };
-      Stream<ConnectionStateUpdate> deviceUpdateStream;
+      Stream<ConnectionStateUpdate?>? deviceUpdateStream;
 
       setUp(() {
         when(_deviceConnector.connect(
@@ -314,7 +322,7 @@ void main() {
       final servicesToDiscover = {
         Uuid.parse('FEFF'): [Uuid.parse('FE1F')]
       };
-      Stream<ConnectionStateUpdate> deviceUpdateStream;
+      Stream<ConnectionStateUpdate?>? deviceUpdateStream;
 
       setUp(() {
         when(_deviceConnector.connectToAdvertisingDevice(
@@ -371,7 +379,7 @@ void main() {
         failure: null,
       );
 
-      Stream<ConnectionStateUpdate> updateStream;
+      Stream<ConnectionStateUpdate>? updateStream;
 
       setUp(() {
         when(_deviceConnector.deviceConnectionStateUpdateStream)
@@ -399,8 +407,8 @@ void main() {
 
       QualifiedCharacteristic char;
 
-      Stream<List<int>> valueStream;
-      Stream<List<int>> resultStream;
+      late Stream<List<int>> valueStream;
+      late Stream<List<int>?> resultStream;
 
       setUp(() {
         char = _createChar();
