@@ -1,35 +1,52 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:flutter_reactive_ble_example/src/ble/ble_device_connector.dart';
+import 'package:flutter_reactive_ble_example/src/ble/ble_device_interactor.dart';
 import 'package:flutter_reactive_ble_example/src/ble/ble_scanner.dart';
 import 'package:flutter_reactive_ble_example/src/ble/ble_status_monitor.dart';
 import 'package:flutter_reactive_ble_example/src/ui/ble_status_screen.dart';
 import 'package:flutter_reactive_ble_example/src/ui/device_list.dart';
 import 'package:provider/provider.dart';
 
+import 'src/ble/ble_logger.dart';
+
 const _themeColor = Colors.lightGreen;
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
 
+  final _bleLogger = BleLogger();
   final _ble = FlutterReactiveBle();
-  final _scanner = BleScanner(_ble);
+  final _scanner = BleScanner(ble: _ble, logMessage: _bleLogger.addToLog);
   final _monitor = BleStatusMonitor(_ble);
-  final _connector = BleDeviceConnector(_ble);
+  final _connector = BleDeviceConnector(
+    ble: _ble,
+    logMessage: _bleLogger.addToLog,
+  );
+  final _serviceDiscoverer = BleDeviceInteractor(
+    bleDiscoverServices: _ble.discoverServices,
+    readCharacteristic: _ble.readCharacteristic,
+    writeWithResponse: _ble.writeCharacteristicWithResponse,
+    writeWithOutResponse: _ble.writeCharacteristicWithoutResponse,
+    subscribeToCharacteristic: _ble.subscribeToCharacteristic,
+    logMessage: _bleLogger.addToLog,
+  );
   runApp(
     MultiProvider(
       providers: [
         Provider.value(value: _scanner),
         Provider.value(value: _monitor),
         Provider.value(value: _connector),
-        StreamProvider<BleScannerState>(
+        Provider.value(value: _serviceDiscoverer),
+        Provider.value(value: _bleLogger),
+        StreamProvider<BleScannerState?>(
           create: (_) => _scanner.state,
           initialData: const BleScannerState(
             discoveredDevices: [],
             scanIsInProgress: false,
           ),
         ),
-        StreamProvider<BleStatus>(
+        StreamProvider<BleStatus?>(
           create: (_) => _monitor.state,
           initialData: BleStatus.unknown,
         ),
@@ -54,12 +71,12 @@ void main() {
 
 class HomeScreen extends StatelessWidget {
   @override
-  Widget build(BuildContext context) => Consumer<BleStatus>(
+  Widget build(BuildContext context) => Consumer<BleStatus?>(
         builder: (_, status, __) {
           if (status == BleStatus.ready) {
             return DeviceListScreen();
           } else {
-            return BleStatusScreen(status: status);
+            return BleStatusScreen(status: status ?? BleStatus.unknown);
           }
         },
       );
