@@ -4,49 +4,52 @@ import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:flutter_reactive_ble_example/src/ble/reactive_state.dart';
 
 class BleDeviceConnector extends ReactiveState<ConnectionStateUpdate> {
-  BleDeviceConnector(this._ble);
+  BleDeviceConnector({
+    required FlutterReactiveBle ble,
+    required Function(String message) logMessage,
+  })   : _ble = ble,
+        _logMessage = logMessage;
 
   final FlutterReactiveBle _ble;
+  final void Function(String message) _logMessage;
 
   @override
   Stream<ConnectionStateUpdate> get state => _deviceConnectionController.stream;
 
   final _deviceConnectionController = StreamController<ConnectionStateUpdate>();
 
-  StreamSubscription<ConnectionStateUpdate> _connection;
+  // ignore: cancel_subscriptions
+  late StreamSubscription<ConnectionStateUpdate> _connection;
 
   Future<void> connect(String deviceId) async {
-    if (_connection != null) {
-      await _connection.cancel();
-    }
+    _logMessage('Start connecting to $deviceId');
     _connection = _ble.connectToDevice(id: deviceId).listen(
-          _deviceConnectionController.add,
-        );
+      (update) {
+        _logMessage(
+            'ConnectionState for device $deviceId : ${update.connectionState}');
+        _deviceConnectionController.add(update);
+      },
+      onError: (e) =>
+          _logMessage('Connecting to device $deviceId resulted in error $e'),
+    );
   }
 
   Future<void> disconnect(String deviceId) async {
-    if (_connection != null) {
-      try {
-        await _connection.cancel();
-      } on Exception catch (e, _) {
-        print("Error disconnecting from a device: $e");
-      } finally {
-        // Since [_connection] subscription is terminated, the "disconnected" state cannot be received and propagated
-        _deviceConnectionController.add(
-          ConnectionStateUpdate(
-            deviceId: deviceId,
-            connectionState: DeviceConnectionState.disconnected,
-            failure: null,
-          ),
-        );
-      }
+    try {
+      _logMessage('disconnecting to device: $deviceId');
+      await _connection.cancel();
+    } on Exception catch (e, _) {
+      _logMessage("Error disconnecting from a device: $e");
+    } finally {
+      // Since [_connection] subscription is terminated, the "disconnected" state cannot be received and propagated
+      _deviceConnectionController.add(
+        ConnectionStateUpdate(
+          deviceId: deviceId,
+          connectionState: DeviceConnectionState.disconnected,
+          failure: null,
+        ),
+      );
     }
-  }
-
-  Future<void> discoverServices(String deviceId) async {
-    await _ble.discoverServices(deviceId).then(
-          (value) => print('Services discovered: $value'),
-        );
   }
 
   Future<void> dispose() async {
