@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import '../model/unit.dart';
+
 /// A disposable resource whose underlying resource can be replaced by another resource
 /// causing automatic disposal of the previous underlying resource.
 class SerialDisposable<T> {
@@ -9,25 +11,24 @@ class SerialDisposable<T> {
     if (_isDisposed) {
       throw _SerialAlreadyDisposed(runtimeType);
     }
-    if (_value != null) await _dispose(_value);
+    if (_value != null) await _dispose(_value!);
     _value = newValue;
-    return newValue;
   }
 
   /// Dispose underlying resource
   Future<void> dispose() async {
     _isDisposed = true;
     if (_value != null) {
-      await _dispose(_value);
+      await _dispose(_value!);
     }
   }
 
   /// Returns whether or not the underlying resource is disposed
   bool get isDisposed => _isDisposed;
 
-  final Future<void> Function(T) _dispose;
+  final Future<Unit> Function(T) _dispose;
   bool _isDisposed = false;
-  T _value;
+  T? _value;
 }
 
 class _SerialAlreadyDisposed extends Error {
@@ -43,5 +44,8 @@ class _SerialAlreadyDisposed extends Error {
 class StreamSubscriptionSerialDisposable
     extends SerialDisposable<StreamSubscription> {
   StreamSubscriptionSerialDisposable()
-      : super((StreamSubscription subscription) => subscription.cancel());
+      : super((StreamSubscription subscription) async {
+          subscription.cancel();
+          return Unit();
+        });
 }
